% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_creation.R
\name{data_2_answers}
\alias{data_2_answers}
\title{Extracts answers from ugly Rds file to human readable table}
\usage{
data_2_answers(filename)
}
\arguments{
\item{filename}{A path to .Rds file generated by exams package}
}
\value{
tibble with answers
}
\description{
Extracts answers from ugly Rds file to human readable table
}
\examples{
# data_2_answers("aaa.Rds")
}
