% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex2RmdExams.R
\name{dollars_to_brackets}
\alias{dollars_to_brackets}
\title{Converts $$...$$ into \[...\]}
\usage{
dollars_to_brackets(string)
}
\arguments{
\item{string}{latex block with $$...$$}
}
\value{
latex block with \[...\]
}
\description{
Converts $$...$$ into \[...\]
}
\details{
Converts $$...$$ into \[...\]
Assumes that the number of $$ is pair.
}
\examples{
string = "$$5 + 5$$ + "
dollars_to_brackets(string)
}
