% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_creation.R
\name{exams2pdf_source}
\alias{exams2pdf_source}
\title{User-friendly version of exams2pdf from exams package}
\usage{
exams2pdf_source(files_sample, n_vars = 1, add_seed = 777,
  pdf_dir = "pdf", language = "ru", tex_dir = "tex",
  name = "the_exam", date = "2018-12-28",
  institution = "Теория вероятностей", logo = "",
  encoding = "UTF-8", samepage = TRUE, reglength = 3, blank = 0,
  template = "plain_no_sweave.tex",
  header = "\\\\input{../header.tex}",
  title = "С Наступающим Новым Годом :)",
  nops = TRUE, shuffle = TRUE)
}
\arguments{
\item{files_sample}{character vector of Rmd file names}

\item{n_vars}{number of variants}

\item{add_seed}{actual seed is variant number plus add_seed}

\item{pdf_dir}{folder for output pdf}

\item{language}{language}

\item{tex_dir}{folder for output tex}

\item{name}{name}

\item{date}{date}

\item{institution}{instituition}

\item{logo}{name of the logo file}

\item{encoding}{encoding}

\item{samepage}{?}

\item{reglength}{?}

\item{blank}{number of blank pages}

\item{template}{name of the template tex file, not used if nops == TRUE}

\item{header}{header tex file name, not used if nops == FALSE}

\item{title}{title}

\item{nops}{logical, whether to use exam2nops or exam2pdf}

\item{shuffle}{logical, whether to shuffle questions in each variant}
}
\value{
Creates tex files and runs latex compilation
}
\description{
User-friendly version of exams2pdf from exams package
}
\examples{
# library(exams)
Sys.which("latexmk") # check whether latex is present and specify path to latexmk
latex_executable <- Sys.which("pdflatex")
options(texi2dvi = latex_executable)
# files_all <- list.files('rmd/', pattern = "*.Rmd", full.names = TRUE)
# exams2pdf_source(files_sample, date = "2019-09-27",
#            n_vars = 2, title = "Теория вероятностей!", institution = "Поехали :)", 
# nops = TRUE, shuffle = TRUE)
}
