% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex2RmdExams.R
\name{extract_from_command}
\alias{extract_from_command}
\title{Extracts unique argument from a latex command.}
\usage{
extract_from_command(element, command = "correctchoice")
}
\arguments{
\item{element}{character string with the command}

\item{command}{command name}
}
\value{
string with the argument of a command
}
\description{
Extracts unique argument from a latex command.
}
\details{
For example, from `\command{aabbcc}` we will extract `aabbcc`.
Assumptions:
1. Unique command with given name.
2. Last closing bracket in element is the bracket we need.
}
\examples{
element = "\n      \\\\correctchoice{ $\\\\frac{S_n - 4n}{10\\\\sqrt{n}}$ }"
extract_from_command(element, command = "correctchoice")
}
