% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta_information}
\alias{get_meta_information}
\title{Get meta information from one or many Rmd rexams files}
\usage{
get_meta_information(file)
}
\arguments{
\item{file}{character vector path to Rmd file or Rmd files}
}
\value{
tibble
}
\description{
Get meta information from one or many Rmd rexams files
}
\examples{
# res = get_meta_information(file = c('1.Rmd', '2.Rmd'))
}
