% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta_information0}
\alias{get_meta_information0}
\title{Get meta information from one Rmd rexams file}
\usage{
get_meta_information0(file)
}
\arguments{
\item{file}{character path to Rmd file}
}
\value{
tibble
}
\description{
Get meta information from one Rmd rexams file
}
\examples{
# res = get_meta_information0(file = '1.Rmd')
}
