% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex2RmdExams.R
\name{global_trim}
\alias{global_trim}
\title{Trim one big string}
\usage{
global_trim(string)
}
\arguments{
\item{string}{original string}
}
\value{
trimmed string
}
\description{
Trim one big string
}
\details{
Trim every space and tab after and before each newline symbol.
}
\examples{
example = "sss\n   dfdfd\n\n ddfdf\n"
global_trim(example)
}
