% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_creation.R
\name{pretemplate2template}
\alias{pretemplate2template}
\title{Transform pretemplate file to template}
\usage{
pretemplate2template(
  path = pretemplate_example("zetex-pre-template.tex"),
  what = c("lhead", "chead", "rhead", "firstpage", "nquestions", "lfoot", "cfoot",
    "rfoot"),
  to = c("Теория вероятностей", "", "1984-01-01", "Удачи", "30",
    "", "Dont panic", "\\\\thepage")
)
}
\arguments{
\item{path}{name of pretemplate}

\item{what}{vector of replaced words in pretemplate}

\item{to}{vector of replacement words}
}
\value{
template as character string
}
\description{
Transform pretemplate file to template.
There are some pretemplates in ints/extdata directory.
}
\examples{
a = pretemplate2template()
}
