% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_creation.R
\name{pretemplate_example}
\alias{pretemplate_example}
\title{Get path to pretemplate examples}
\usage{
pretemplate_example(path = NULL)
}
\arguments{
\item{path}{name of pretemplate}
}
\value{
vector of file names of path to specific file
}
\description{
Get path to pretemplate examples. 
There are some pretemplates in inst/extdata directory.
}
\examples{
pretemplate_example()
pretemplate_example('zetex-pre-template.tex')
}
