% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_creation.R
\name{pretemplate_params}
\alias{pretemplate_params}
\title{Get a list of pretemplate parameters}
\usage{
pretemplate_params(path = pretemplate_example("zetex-pre-template.tex"))
}
\arguments{
\item{path}{name of pretemplate}
}
\value{
character vector of pretemplate params
}
\description{
Get a list of fields that can be replaced by something meaningful. 
Use four slashes to write one slash.
}
\examples{
pretemplate_params()
}
