% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdExams2beamerlatex.R
\name{RmdExams2beamerlatex}
\alias{RmdExams2beamerlatex}
\title{Converts a folder of .Rmd files (rexams package formatted) into a Beamer presentation.}
\usage{
RmdExams2beamerlatex(path2RmdFilesFolder)
}
\arguments{
\item{path2RmdFilesFolder}{folder name}
}
\value{
Creates a tex file for beamer presentation
}
\description{
Converts a folder of .Rmd files (rexams package formatted) into a Beamer presentation.
}
\examples{
# latex2RmdExams("2015_midterm.txt")
}
