% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmdExams2latex.R
\name{RmdExams2latex}
\alias{RmdExams2latex}
\title{Convert a folder of Rmd files into the one Latex document.}
\usage{
RmdExams2latex(path2RmdFilesFolder)
}
\arguments{
\item{path2RmdFilesFolder}{A path to the folder containing the files. The last element of the path must be the name of the folder.}
}
\value{
Creates an 'exam.txt' file in the folder with the Rmd files.
}
\description{
Convert a folder of Rmd files into the one Latex document.
}
\examples{
RmdExams2latex("2015_midterm")
}
