% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_creation.R
\name{data_2_answers}
\alias{data_2_answers}
\title{Extracts answers from ugly list to human readable table}
\usage{
data_2_answers(
  exam_data,
  preserved_fields = c("name", "Language", "Level", "Type")
)
}
\arguments{
\item{exam_data}{ugly list with answers from R exams package}

\item{preserved_fields}{preserved fields in metainfo}
}
\value{
tibble with answers
}
\description{
Extracts answers from ugly list to human readable table
}
\examples{
# data_2_answers(exam_data)
}
