% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_creation.R
\name{pretemplate2template}
\alias{pretemplate2template}
\title{Transform pretemplate file to template}
\usage{
pretemplate2template(path = pretemplate_example("zetex-pre-template.tex"), ...)
}
\arguments{
\item{path}{path to pretemplate}

\item{...}{fields replaced in pretemplate as character vectors.
The length of each character vector should be 1 or equal to the number of variants.}
}
\value{
templates as character vector
}
\description{
Transform pretemplate file to template.
There are some pretemplates in inst/extdata directory.
}
\examples{
template_text = pretemplate2template(
  lhead = 'Теория вероятностей', chead = '', rhead = '1984-01-01', 
  firstpage = 'Удачи!', nquestions = '30', 
  lfoot = c('Вариант 1', 'Вариант 2'), 
  cfoot = 'Don\'t panic', rfoot = '\\\\\\\\thepage')
cat(template_text[1])
}
